requirements.txt
astrapy==1.1.0
python-dotenv==1.0.1
pandas==2.2.2